feat: two-tier sync + identity + outcome loop + gerencial dashboard
This session's work (all green: 1443 passed/1 skip, ruff+mypy clean):

- Two-tier sync (sync_git.py, identity.py): explicit LOCAL (intra-machine,
  shared sqlite/WAL, no git) vs REMOTE (git memo-sync) models. sync_once()
  is a single machine-level coordinator — flock on .sync.lock + pull-rebase-
  before-push. Triggers: `sync auto` (debounced, per-prompt) + `sync once`
  (Stop flush). Stranded-commit retry + sync_pending marker. New identity
  layer (stable machine_id = device_id, hostname, session, terminal); commits
  attributed [hostname·session].

- Durable capture (capture.py, cli_capture.py): `memo capture-tick` —
  per-session watermark, captures only new turns mid-session so a long
  session's insight reaches .md before Stop (MEMO_CAPTURE_INTERVAL_S).

- The Outcome Loop (outcome.py, cli_outcome.py): roi_score driven by real
  grounding outcomes (not mere access) — reconcile_roi + `memo gaps` +
  dead-weight archival, gated by MEMO_OUTCOME_RANKING_ENABLED.

- Gerencial dashboard (web/build.py): rebuilt around "is memo useful as
  primary memory?" — verdict, consult funnel, per-tool readers, tokens-saved
  detail, knowledge-gaps + sync-health panels.

- Tokens-saved metric (cli_roi.py): estimate model tokens saved per grounded
  recall / re-ask avoided.

- Trinity attribution: MCP/CLI consults fall back to MEMO_SOURCE env +
  clientInfo.name; windsurf retired → devin-desktop in EXPECTED_CONSUMERS.

- CLAUDE.md sync section + docs/superpowers/specs/ design doc.

Also sweeps in pre-existing uncommitted working-tree changes present at
session start (embedder/llm/ingest/repo-index/server churn) — the full suite
passes with them in place.

+++ src/memo/flags_ingest.py
@@ -62,7 +62,7 @@ SPECS: tuple[FlagSpec, ...] = (
     ),
     _spec("MEMO_REPO_FLUSH_BATCH", "int", None, "repo", "Rows per DB flush during repo index."),
     _spec(
-        "MEMO_REPO_GIT_TIMEOUT_S", "int", None, "repo", "Timeout (s) for git operations on clone."
+        "MEMO_REPO_GIT_TIMEOUT_S", "float", None, "repo", "Timeout (s) for git operations on clone."
     ),
     # embedder daemon / client
     _spec(
